requirements/dev.txt
-r common.txt

# Sphinx depends on Pygments and will install Pygments if it is not already
# installed in your environment. The latest release of Pygments 2.1.3
# (released Mar 2, 2016) will cause Sphinx to fail in compiling a regex,
# which will eventually cause fab build to fail. The most recent compatible
# version of Pygments is 2.1; until a compatible version of Pygments is
# released, please use Pygments 2.1.

# At the time of writing the compatible version of Sphinx is 1.4.4.

Pygments==2.1
Sphinx
coverage
mock
django-nose>=1.4
fabric
graphviz
django-debug-toolbar